services/gemini-live-tts/requirements.txt
google-genai>=0.8.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
websockets>=12.0
